Move supports-method from implementing validator to AbstractStatementValidator class
