+++ webauthn4j-core/src/main/java/com/webauthn4j/validator/attestation/statement/androidkey/AndroidKeyAttestationStatementValidator.java
@@ -17,12 +17,11 @@
 package com.webauthn4j.validator.attestation.statement.androidkey;
 
 import com.webauthn4j.response.attestation.statement.AndroidKeyAttestationStatement;
-import com.webauthn4j.response.attestation.statement.AttestationStatement;
 import com.webauthn4j.response.attestation.statement.AttestationType;
 import com.webauthn4j.util.MessageDigestUtil;
 import com.webauthn4j.util.SignatureUtil;
 import com.webauthn4j.validator.RegistrationObject;
-import com.webauthn4j.validator.attestation.statement.AttestationStatementValidator;
+import com.webauthn4j.validator.attestation.statement.AbstractStatementValidator;
 import com.webauthn4j.validator.exception.BadAttestationStatementException;
 import com.webauthn4j.validator.exception.BadSignatureException;
 import com.webauthn4j.validator.exception.PublicKeyMismatchException;
@@ -31,7 +30,7 @@ import java.nio.ByteBuffer;
 import java.security.*;
 import java.security.cert.Certificate;
 
-public class AndroidKeyAttestationStatementValidator implements AttestationStatementValidator {
+public class AndroidKeyAttestationStatementValidator extends AbstractStatementValidator<AndroidKeyAttestationStatement> {
 
     // ~ Instance fields
     // ================================================================================================
@@ -68,13 +67,6 @@ public class AndroidKeyAttestationStatementValidator implements AttestationState
         keyDescriptionValidator.validate(attestationStatement.getX5c().getEndEntityAttestationCertificate().getCertificate(), clientDataHash, teeEnforcedOnly);
 
         return AttestationType.BASIC;
-    }
-
-
-    @Override
-    public boolean supports(RegistrationObject registrationObject) {
-        AttestationStatement attestationStatement = registrationObject.getAttestationObject().getAttestationStatement();
-        return AndroidKeyAttestationStatement.class.isAssignableFrom(attestationStatement.getClass());
     }
 
     private void validateSignature(RegistrationObject registrationObject) {
