+++ webauthn4j-core/src/main/java/com/webauthn4j/validator/attestation/statement/androidsafetynet/AndroidSafetyNetAttestationStatementValidator.java
@@ -16,11 +16,14 @@
 
 package com.webauthn4j.validator.attestation.statement.androidsafetynet;
 
-import com.webauthn4j.response.attestation.statement.*;
+import com.webauthn4j.response.attestation.statement.AndroidSafetyNetAttestationStatement;
+import com.webauthn4j.response.attestation.statement.AttestationCertificate;
+import com.webauthn4j.response.attestation.statement.AttestationType;
+import com.webauthn4j.response.attestation.statement.Response;
 import com.webauthn4j.util.Base64Util;
 import com.webauthn4j.util.MessageDigestUtil;
 import com.webauthn4j.validator.RegistrationObject;
-import com.webauthn4j.validator.attestation.statement.AttestationStatementValidator;
+import com.webauthn4j.validator.attestation.statement.AbstractStatementValidator;
 import com.webauthn4j.validator.exception.BadAttestationStatementException;
 import com.webauthn4j.validator.exception.BadSignatureException;
 
@@ -31,7 +34,7 @@ import java.time.ZoneOffset;
 import java.util.Arrays;
 import java.util.Objects;
 
-public class AndroidSafetyNetAttestationStatementValidator implements AttestationStatementValidator {
+public class AndroidSafetyNetAttestationStatementValidator extends AbstractStatementValidator<AndroidSafetyNetAttestationStatement> {
 
     // ~ Instance fields
     // ================================================================================================
@@ -107,12 +110,6 @@ public class AndroidSafetyNetAttestationStatementValidator implements Attestatio
         if (!Arrays.equals(hash, Base64Util.decode(nonce))) {
             throw new BadAttestationStatementException("Nonce doesn't match");
         }
-    }
-
-    @Override
-    public boolean supports(RegistrationObject registrationObject) {
-        AttestationStatement attestationStatement = registrationObject.getAttestationObject().getAttestationStatement();
-        return AndroidSafetyNetAttestationStatement.class.isAssignableFrom(attestationStatement.getClass());
     }
 
     public int getForwardThreshold() {
