+++ webauthn4j-core/src/main/java/com/webauthn4j/validator/attestation/statement/tpm/NullTPMAttestationStatementValidator.java
@@ -16,13 +16,12 @@
 
 package com.webauthn4j.validator.attestation.statement.tpm;
 
-import com.webauthn4j.response.attestation.statement.AttestationStatement;
 import com.webauthn4j.response.attestation.statement.AttestationType;
 import com.webauthn4j.response.attestation.statement.TPMAttestationStatement;
 import com.webauthn4j.validator.RegistrationObject;
-import com.webauthn4j.validator.attestation.statement.AttestationStatementValidator;
+import com.webauthn4j.validator.attestation.statement.AbstractStatementValidator;
 
-public class NullTPMAttestationStatementValidator implements AttestationStatementValidator {
+public class NullTPMAttestationStatementValidator extends AbstractStatementValidator<TPMAttestationStatement> {
     @Override
     public AttestationType validate(RegistrationObject registrationObject) {
         if (!supports(registrationObject)) {
@@ -30,11 +29,5 @@ public class NullTPMAttestationStatementValidator implements AttestationStatemen
         }
 
         return AttestationType.NONE;
-    }
-
-    @Override
-    public boolean supports(RegistrationObject registrationObject) {
-        AttestationStatement attestationStatement = registrationObject.getAttestationObject().getAttestationStatement();
-        return TPMAttestationStatement.class.isAssignableFrom(attestationStatement.getClass());
     }
 }
